Fixes how backspace works at beginning of blocks.
