+++ edit-client/src/walkers.rs
@@ -270,6 +270,36 @@ impl<'a> Walker<'a> {
         matched
     }
 
+    pub fn at_start_of_block(&mut self) -> bool {
+        let mut matched = false;
+        take_mut::take(&mut self.stepper, |prev_stepper| {
+            let mut rstepper = prev_stepper.clone().rev();
+
+            // Iterate until we reach a block.
+            matched = loop {
+                if rstepper.next().is_none() {
+                    break false;
+                }
+                match rstepper.doc.head() {
+                    Some(DocGroup(attrs, _)) => {
+                        if is_any_caret(&attrs) {
+                            continue;
+                        }
+                        if is_block(&attrs) {
+                            break true;
+                        }
+                    }
+                    _ => {}
+                }
+                break false;
+            };
+
+            prev_stepper
+        });
+
+        matched
+    }
+
     pub fn back_block(&mut self) -> bool {
         let mut matched = false;
         take_mut::take(&mut self.stepper, |prev_stepper| {
